Small update Saved Message Overalay
Updating all files to support a new method for showing saved settings... rather than use a basic show message we are now using an overlay which shows the developer that settings have been saved. The overlay will show up in the center of the IDE and fade away within a few secconds.